test: 新增Parse-SDK-JS测试用例和demo
